+++ src/lib/components/filter/FilterDropdownItem.tsx
@@ -2,6 +2,8 @@ import { Flex, ListItem, Text } from "@chakra-ui/react";
 import type { CSSProperties, ReactNode } from "react";
 
 import { CustomIcon } from "../icon";
+import { AmpEvent, trackUseFilter } from "lib/amplitude";
+import { toggleItem } from "lib/utils";
 
 const listItemProps: CSSProperties = {
   borderRadius: "8px",
@@ -11,27 +13,64 @@ const listItemProps: CSSProperties = {
 };
 
 interface FilterDropdownItemProps {
-  isOptionSelected: boolean;
+  result: string | string[] | undefined;
+  option: string;
   filterDropdownComponent: ReactNode;
-  onSelect: () => void;
+  setIsDropdown: (isDropdown: boolean) => void;
+  setKeyword: (keyword: string) => void;
+  setResult: (option: string | string[] | undefined) => void;
 }
 
 export const FilterDropdownItem = ({
-  isOptionSelected,
+  result,
+  option,
   filterDropdownComponent,
-  onSelect,
-}: FilterDropdownItemProps) => (
-  <ListItem
-    style={listItemProps}
-    _hover={{ bg: "gray.800" }}
-    transition="all 0.25s ease-in-out"
-    onClick={onSelect}
-  >
-    <Flex alignItems="center" justifyContent="space-between">
-      <Text wordBreak="break-all" lineHeight="1.2">
-        {filterDropdownComponent}
-      </Text>
-      {isOptionSelected && <CustomIcon name="check" color="gray.600" />}
-    </Flex>
-  </ListItem>
-);
+  setIsDropdown,
+  setKeyword,
+  setResult,
+}: FilterDropdownItemProps) => {
+  const isOptionSelected =
+    typeof result === "string" ? result === option : result?.includes(option);
+
+  const selectOption = () => {
+    setKeyword("");
+
+    if (typeof result === "string" || !result || !option) {
+      setIsDropdown(false);
+
+      if (typeof result === "string") {
+        trackUseFilter(AmpEvent.USE_FILTER_PROPOSALS_STATUS, result, "add");
+        setResult(option);
+        return;
+      }
+
+      trackUseFilter(AmpEvent.USE_FILTER_PROPOSALS_STATUS, result, "remove");
+      setResult(option);
+      return;
+    }
+
+    if (result.includes(option)) {
+      trackUseFilter(AmpEvent.USE_FILTER_PROPOSALS_STATUS, result, "remove");
+    } else {
+      trackUseFilter(AmpEvent.USE_FILTER_PROPOSALS_STATUS, result, "add");
+    }
+
+    setResult(toggleItem(result, option));
+  };
+
+  return (
+    <ListItem
+      style={listItemProps}
+      _hover={{ bg: "gray.800" }}
+      transition="all 0.25s ease-in-out"
+      onClick={selectOption}
+    >
+      <Flex alignItems="center" justifyContent="space-between">
+        <Text wordBreak="break-all" lineHeight="1.2">
+          {filterDropdownComponent}
+        </Text>
+        {isOptionSelected && <CustomIcon name="check" color="gray.600" />}
+      </Flex>
+    </ListItem>
+  );
+};