+++ src/lib/pages/proposals/components/ProposalTypeFilter.tsx
@@ -68,9 +68,6 @@ export const ProposalTypeFilter = forwardRef<
 
     if (!proposalTypes) return null;
 
-    const isOptionSelected = (option: ProposalType) =>
-      result.some((selectedOption) => selectedOption === option);
-
     const selectOption = (option: ProposalType) => {
       if (inputRef.current) {
         setKeyword("");
@@ -128,8 +125,11 @@ export const ProposalTypeFilter = forwardRef<
               <FilterDropdownItem
                 key={cosmosType}
                 filterDropdownComponent={cosmosType}
-                isOptionSelected={isOptionSelected(cosmosType)}
-                onSelect={() => selectOption(cosmosType)}
+                result={result}
+                option={cosmosType}
+                setIsDropdown={setIsDropdown}
+                setKeyword={setKeyword}
+                setResult={(opt) => setResult(opt as ProposalType[])}
               />
             ))}
             {cosmosTypes.length && <Divider borderColor="gray.700" />}
@@ -138,8 +138,11 @@ export const ProposalTypeFilter = forwardRef<
               <FilterDropdownItem
                 key={generalType}
                 filterDropdownComponent={generalType}
-                isOptionSelected={isOptionSelected(generalType)}
-                onSelect={() => selectOption(generalType)}
+                result={result}
+                option={generalType}
+                setIsDropdown={setIsDropdown}
+                setKeyword={setKeyword}
+                setResult={(opt) => setResult(opt as ProposalType[])}
               />
             ))}
           </DropdownContainer>