+++ src/lib/services/tx/lcd.ts
@@ -1,10 +1,6 @@
 import axios from "axios";
 
-import type { Option } from "lib/types";
-
-export const getTxDataLcd = async (endpoint: string, txHash: Option<string>) =>
+export const getTxDataLcd = async (endpoint: string, txHash: string) =>
   axios
-    .get(
-      `${endpoint}/cosmos/tx/v1beta1/txs/${encodeURIComponent(txHash ?? "")}`
-    )
+    .get(`${endpoint}/cosmos/tx/v1beta1/txs/${encodeURIComponent(txHash)}`)
     .then(({ data }) => data);