+++ src/lib/types/proposal.ts
@@ -22,16 +22,6 @@ export enum ProposalStatus {
 }
 export const zProposalStatus = z.nativeEnum(ProposalStatus);
 
-export enum ProposalStatusLcd {
-  ALL = "ALL_STATUSES",
-  PROPOSAL_STATUS_DEPOSIT_PERIOD = "PROPOSAL_STATUS_DEPOSIT_PERIOD",
-  PROPOSAL_STATUS_VOTING_PERIOD = "PROPOSAL_STATUS_VOTING_PERIOD",
-  PROPOSAL_STATUS_PASSED = "PROPOSAL_STATUS_PASSED",
-  PROPOSAL_STATUS_REJECTED = "PROPOSAL_STATUS_REJECTED",
-  PROPOSAL_STATUS_FAILED = "PROPOSAL_STATUS_FAILED",
-}
-export const zProposalStatusLcd = z.nativeEnum(ProposalStatusLcd);
-
 export enum ProposalTypeCosmos {
   TEXT = "Text",
   PARAMETER = "ParameterChange",